// src/app/models/user.ts
export class User {
    id: number;
    name: string;
    email: string;
    contactno: number;
    password: string;
    shippingAddress: string;
    landMark: string;
    shippingState: string;
    shippingCity: string;
    shippingPincode: number;
    billingAddress: string;
    billingState: string;
    billingCity: string;
    billingPincode: number;
    regDate: string;
    updationDate: string;
    ip: any;
    message: string;
}